feat: trigger prompt volume analysis after checkout
