+++ server/src/routes/volumes.js
@@ -1,6 +1,5 @@
 import { Router } from 'express';
 import { resolveModel } from '../lib/ai-provider.js';
-import { getSearchVolumes } from '../lib/dataforseo.js';
 import { regionToLocationCode, languageToCode } from '../lib/dataforseo-codes.js';
 import {
   requireFeature,
@@ -10,87 +9,11 @@ import {
 } from '../lib/plan-guard.js';
 import supabaseAdmin from '../config/supabase.js';
 import { assertBrandAccess, assertPromptAccess } from '../lib/access.js';
-import { AI_VOLUME_MULTIPLIER, extractIntentKeywords } from '../lib/intent-extraction.js';
+import { extractIntentKeywords } from '../lib/intent-extraction.js';
+import { mapVolumeRow, fetchAndSaveVolumes } from '../lib/volume-analysis.js';
+import { analyzeBrandVolumes } from '../lib/volume-analysis.js';
 
 const router = Router();
-
-function mapVolumeRow(saved) {
-  return {
-    id: saved.id,
-    promptId: saved.prompt_id,
-    intent: saved.intent,
-    keywords: saved.keywords,
-    googleVolumes: saved.google_volumes,
-    totalGoogleVolume: saved.total_google_volume,
-    aiVolumeMultiplier: parseFloat(saved.ai_volume_multiplier),
-    estAiVolume: saved.est_ai_volume,
-    competitionIndex: saved.competition_index ?? null,
-    competition: saved.competition ?? null,
-    locationCode: saved.location_code,
-    languageCode: saved.language_code,
-    fetchedAt: saved.fetched_at,
-  };
-}
-
-async function fetchAndSaveVolumes(promptId, keywords, intent, locationCode, languageCode) {
-  const volumes = await getSearchVolumes(keywords, {
-    locationCode: locationCode || undefined,
-    languageCode: languageCode || undefined,
-  });
-
-  // google_volumes stays a { keyword: number } map for the UI. Competition is
-  // aggregated per prompt as a volume-weighted average of the keyword indices.
-  const googleVolumes = {};
-  let totalGoogleVolume = 0;
-  let competitionWeightedSum = 0;
-  let competitionWeight = 0;
-  for (const [keyword, data] of Object.entries(volumes)) {
-    googleVolumes[keyword] = data.volume;
-    totalGoogleVolume += data.volume;
-    if (data.competitionIndex !== null && data.competitionIndex !== undefined) {
-      competitionWeightedSum += data.competitionIndex * data.volume;
-      competitionWeight += data.volume;
-    }
-  }
-
-  const competitionIndex =
-    competitionWeight > 0 ? Math.round(competitionWeightedSum / competitionWeight) : null;
-  const competition =
-    competitionIndex === null
-      ? null
-      : competitionIndex <= 33
-        ? 'LOW'
-        : competitionIndex <= 66
-          ? 'MEDIUM'
-          : 'HIGH';
-
-  const estAiVolume = Math.round(totalGoogleVolume * AI_VOLUME_MULTIPLIER);
-
-  const { data: saved, error: dbError } = await supabaseAdmin
-    .from('prompt_volumes')
-    .upsert(
-      {
-        prompt_id: promptId,
-        intent,
-        keywords,
-        google_volumes: googleVolumes,
-        total_google_volume: totalGoogleVolume,
-        ai_volume_multiplier: AI_VOLUME_MULTIPLIER,
-        est_ai_volume: estAiVolume,
-        competition_index: competitionIndex,
-        competition,
-        location_code: locationCode || null,
-        language_code: languageCode || null,
-        fetched_at: new Date().toISOString(),
-      },
-      { onConflict: 'prompt_id' },
-    )
-    .select()
-    .single();
-
-  if (dbError) throw new Error(dbError.message);
-  return saved;
-}
 
 /**
  * POST /api/volumes/analyze
@@ -195,7 +118,7 @@ router.post('/analyze-batch', requireFeature('prompt_volumes'), async (req, res)
   try {
     const { remaining, orgId } = await enforceVolumeQuota(req.user.id);
 
-    const { prompts, locationCode, languageCode, model, force } = req.body;
+    const { prompts, force } = req.body;
 
     if (!Array.isArray(prompts) || prompts.length === 0) {
       return res.status(400).json({ error: 'prompts array is required and must not be empty' });
@@ -209,79 +132,25 @@ router.post('/analyze-batch', requireFeature('prompt_volumes'), async (req, res)
 
     await assertPromptAccess(promptIds, req.user.id);
 
-    // Resolve DataForSEO location/language from the brand the prompts belong to,
-    // unless the caller explicitly passed an override in the request body.
-    let resolvedLocationCode = locationCode;
-    let resolvedLanguageCode = languageCode;
-    if (resolvedLocationCode == null || resolvedLanguageCode == null) {
-      const { data: brandRow } = await supabaseAdmin
-        .from('prompts')
-        .select('prompt_sets!inner(brands!inner(region, language))')
-        .in('id', promptIds)
-        .limit(1)
-        .maybeSingle();
-      const brand = brandRow?.prompt_sets?.brands;
-      if (brand) {
-        if (resolvedLocationCode == null) {
-          resolvedLocationCode = regionToLocationCode(brand.region);
-        }
-        if (resolvedLanguageCode == null) {
-          resolvedLanguageCode = languageToCode(brand.language);
-        }
-      }
+    const { data: brandRow, error: brandError } = await supabaseAdmin
+      .from('prompts')
+      .select('prompt_sets!inner(brands!inner(id))')
+      .in('id', promptIds)
+      .limit(1)
+      .maybeSingle();
+
+    if (brandError) {
+      throw new Error(`Failed to resolve brand: ${brandError.message}`);
+    }
+    const brandId = brandRow?.prompt_sets?.brands?.id;
+
+    if (!brandId) {
+      return res.status(400).json({
+        error: 'Unable to resolve brand from prompts',
+      });
     }
 
-    let existingMap = {};
-
-    if (!force) {
-      const { data: existingRows } = await supabaseAdmin
-        .from('prompt_volumes')
-        .select('prompt_id, intent, keywords')
-        .in('prompt_id', promptIds);
-
-      if (existingRows) {
-        for (const row of existingRows) {
-          if (row.keywords?.length) {
-            existingMap[row.prompt_id] = {
-              intent: row.intent,
-              keywords: row.keywords,
-            };
-          }
-        }
-      }
-    }
-
-    const aiModel = resolveModel(model);
-    const results = [];
-
-    for (const { promptId, promptText } of prompts) {
-      try {
-        let intent;
-        let keywords;
-
-        const cached = existingMap[promptId];
-        if (cached) {
-          intent = cached.intent;
-          keywords = cached.keywords;
-        } else {
-          const intentResult = await extractIntentKeywords(promptText, aiModel);
-          intent = intentResult.intent;
-          keywords = intentResult.keywords;
-        }
-
-        const saved = await fetchAndSaveVolumes(
-          promptId,
-          keywords,
-          intent,
-          resolvedLocationCode,
-          resolvedLanguageCode,
-        );
-        results.push(mapVolumeRow(saved));
-      } catch (err) {
-        results.push({ promptId, error: err.message });
-      }
-    }
-
+    const results = await analyzeBrandVolumes(brandId, { force });
     const successCount = results.filter((r) => !r.error).length;
     if (successCount > 0 && orgId) {
       await supabaseAdmin.from('volume_usage').insert({
